% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{g_of_hp}
\alias{g_of_hp}
\title{g of hp function}
\usage{
g_of_hp(c, h, d)
}
\arguments{
\item{c}{this is a constant}

\item{h}{this is a hare}

\item{d}{this is something}
}
\value{
a math value
}
\description{
g of hp function
}
