% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello function}
\usage{
hello(x, y, z)
}
\arguments{
\item{x}{this is the hare}

\item{y}{this is the lynx}

\item{z}{there is no z}
}
\value{
data.frame
}
\description{
hello function
}
\examples{
hello(hi, hi, hi)
}
