% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulation of Predator and Prey Dynamics}
\usage{
sim(H0 = 2, P0 = 2, a = 0.01, b = 0.01, c = 0.01, d = 0.01,
  t = 1000)
}
\arguments{
\item{H0}{Hare initial population}

\item{P0}{Lynx initial population}

\item{a}{Hare intercept parameter}

\item{b}{Hare slope parameter}

\item{c}{Lynx intercept parameter}

\item{d}{Lynx slope parameter}

\item{t}{Time (number of iterations)}
}
\value{
data.frame of hare and lynx populations
}
\description{
Simulation of Predator and Prey Dynamics
}
