% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{simulation_one}
\alias{simulation_one}
\title{simulation function}
\usage{
simulation_one(t, x, y, a, b, c, d)
}
\arguments{
\item{t}{is time}

\item{x}{is hare}

\item{y}{is lynx}

\item{a}{is some constant}

\item{b}{is some constant}

\item{c}{is some other constant}

\item{d}{is some other constant}
}
\value{
returns a data frame
}
\description{
simulation function
}
\examples{

}
